requirements.txt
crewai
langchain
langchain-groq
python-dotenv
